+++ frontend/src/pages/MainPage/index.js
@@ -1,5 +1,5 @@
 import BannerList from '../../components/Banner/BannerList';
-import bannerList from './bannerList';
+import bannerList from '../../configs/bannerList';
 
 const MainPage = () => {
   return (
